Correção de bugs, inicio da edição da agenda

+++ ServicoAkiPrestador/app/src/main/java/br/com/justworks/prestador/ServicoAki/Model/Address.java
@@ -119,7 +119,7 @@ public class Address {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
@@ -127,7 +127,7 @@ public class Address {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
